requirements.txt
plyfile
tqdm
dearpygui==1.11.1
opencv-python
pillow
scikit-learn
joblib
numpy==1.24.4
open3d==0.16.0
matplotlib
trimesh
lpips
e3nn
einops
einsum
kornia
pyrender

-e submodules/diff-surfel-rasterization
submodules/simple-knn
